pyproject.toml
[project]
name = "event-driven-automation"
version = "0.1.0"
description = "Curso para Semana de Capacitação 11 - NIC.br"
authors = [
    { name = "Adilson Torres", email = "adilson@nic.br" },
    { name = "William Prado", email = "wsprado@nic.br" }
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.116.1",
    "requests>=2.32.5",
    "temporalio>=1.16.0",
    "uvicorn>=0.35.0",
    "napalm>=5.1.0",
    "netmiko>=4.6.0",
    "nornir>=3.5.0",
    "nornir_napalm>=0.5.0",
    "nornir-netbox>=0.3.0",
    "nornir-netmiko>=1.0.1",
    "nornir-utils>=0.2.0",
    "python-dotenv>=1.1.1",
    "pynetbox>=7.5.0",
    "pyeapi>=1.0.4"
]
